[SWM-62] 유저 분석 페이지 출력

+++ src/main/java/com/swmStrong/demo/domain/usageLog/dto/SaveUsageLogDto.java
@@ -6,7 +6,7 @@ public record SaveUsageLogDto(
         String userId,
         String title,
         String app,
-        long duration,
+        double duration,
         LocalDateTime timestamp
 ) {
 }
