+++ src/main/java/com/swmStrong/demo/domain/usageLog/dto/UsageLogResponseDto.java
@@ -8,7 +8,7 @@ public record UsageLogResponseDto(
         String userId,
         String title,
         String app,
-        long duration,
+        double duration,
         LocalDateTime timestamp
 ) {
     public static UsageLogResponseDto from(UsageLog usageLog) {
